More refactoring
- PackInfo -> PackInformation
- Establish that `Pack` is engine-scoped. It is not the intention of
this API to instill that client resources should be loaded server side
nor should server data be loaded client side. That does not mean that a
plugin couldn't take it upon themselves to do so (they are simply files
after all) but it is the goal of the API to correctly establish what a
developer *should* be doing ultimately.
- Per the former, move retrieval of a PluginContainer's pack to
ResourceManager
- Migrate PackVersion/PackType to enums. There is no reason a developer
should be creating additional versions or types via the API.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>